remove gcc-4.8.1 warnings. fix crash in summarize

+++ src/benchbitpacking.cpp
@@ -24,32 +24,32 @@ void maskfnc(vector<uint32_t, cacheallocator> & out, const uint32_t L) {
 
 void fastpack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         fastpack(&data[0] + 32 * k, &out[0] + bit * k, bit);
     }
 }
 
 void fastpackwithoutmask(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         fastpackwithoutmask(&data[0] + 32 * k, &out[0] + bit * k, bit);
     }
 }
 
 void fastunpack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = out.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = out.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         fastunpack(&data[0] + bit * k, &out[0] + 32 * k, bit);
     }
 }
 
 void simdpack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 128; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 128; ++k) {
         SIMD_fastpack_32(&data[0] + 128 * k,
                 reinterpret_cast<__m128i *> (&out[0] + 4 * bit * k), bit);
     }
@@ -57,8 +57,8 @@ void simdpack(const vector<uint32_t, cacheallocator> & data,
 
 void simdpackwithoutmask(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 128; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 128; ++k) {
         SIMD_fastpackwithoutmask_32(&data[0] + 128 * k,
                 reinterpret_cast<__m128i *> (&out[0] + 4 * bit * k), bit);
     }
@@ -66,8 +66,8 @@ void simdpackwithoutmask(const vector<uint32_t, cacheallocator> & data,
 
 void simdunpack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = out.size();
-    for (uint32_t k = 0; k < N / 128; ++k) {
+    const size_t N = out.size();
+    for (size_t k = 0; k < N / 128; ++k) {
         SIMD_fastunpack_32(
                 reinterpret_cast<const __m128i *> (&data[0] + 4 * bit * k),
                 &out[0] + 128 * k, bit);
@@ -77,8 +77,8 @@ void simdunpack(const vector<uint32_t, cacheallocator> & data,
 
 void horizontalunpack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = out.size();
-    for (uint32_t k = 0; k < N / 128; ++k) {
+    const size_t N = out.size();
+    for (size_t k = 0; k < N / 128; ++k) {
         simdhunpack(
                 reinterpret_cast<const uint8_t *> (&data[0] + 4 * bit * k),
                 &out[0] + 128 * k, bit);
@@ -87,48 +87,48 @@ void horizontalunpack(const vector<uint32_t, cacheallocator> & data,
 
 void pack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         pack<true> (&data[0] + 32 * k, &out[0] + bit * k, bit);
     }
 }
 
 void packwithoutmask(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         pack<false> (&data[0] + 32 * k, &out[0] + bit * k, bit);
     }
 }
 
 void pack_tight(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         pack_tight<true> (&data[0] + 32 * k, &out[0] + bit * k, bit);
     }
 }
 
 void pack_tightwithoutmask(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = data.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = data.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         pack_tight<false> (&data[0] + 32 * k, &out[0] + bit * k, bit);
     }
 }
 
 void unpack_tight(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = out.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = out.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         unpack_tight(&data[0] + bit * k, &out[0] + 32 * k, bit);
     }
 }
 
 void unpack(const vector<uint32_t, cacheallocator> & data,
         vector<uint32_t, cacheallocator> & out, const uint32_t bit) {
-    const uint32_t N = out.size();
-    for (uint32_t k = 0; k < N / 32; ++k) {
+    const size_t N = out.size();
+    for (size_t k = 0; k < N / 32; ++k) {
         unpack(&data[0] + bit * k, &out[0] + 32 * k, bit);
     }
 }
@@ -140,7 +140,7 @@ bool equalOnFirstBits(const container32bit & data,
     if (bit == 32) {
         return data == recovered;
     }
-    for (uint32_t k = 0; k < data.size(); ++k) {
+    for (size_t k = 0; k < data.size(); ++k) {
         if (data[k] % (1U << bit) != recovered[k] % (1U << bit)) {
             cout << " They differ at k = " << k << " data[k]= " << data[k]
                     << " recovered[k]=" << recovered[k] << endl;
@@ -155,9 +155,9 @@ void simplebenchmark(uint32_t N = 1U << 16, uint32_t T = 1U << 9) {
     vector<uint32_t, cacheallocator> compressed(N, 0);
     vector<uint32_t, cacheallocator> recovered(N, 0);
     WallClockTimer z;
-    double packtime, packtimewm, unpacktime;
-    double simdpacktime, simdpacktimewm, simdunpacktime;
-    double horizontalunpacktime;
+    uint64_t packtime, packtimewm, unpacktime;
+    uint64_t simdpacktime, simdpacktimewm, simdunpacktime;
+    uint64_t horizontalunpacktime;
 
     cout << "#million of integers per second: higher is better" << endl;
     cout << "#bit, pack, pack without mask, unpack" << endl;
@@ -248,15 +248,15 @@ void simplebenchmark(uint32_t N = 1U << 16, uint32_t T = 1U << 9) {
             }
 
             cout << std::setprecision(4) << bit << "\t\t" << N * (T - 1)
-                    / (packtime) << "\t\t" << N * (T - 1) / (packtimewm)
-                    << "\t\t\t" << N * (T - 1) / (unpacktime) << "\t\t";
+                    / double(packtime) << "\t\t" << N * (T - 1) / double(packtimewm)
+                    << "\t\t\t" << N * (T - 1) / double(unpacktime) << "\t\t";
 
             cout << std::setprecision(4) << bit << "\t\t" << N * (T - 1)
-                    / (simdpacktime) << "\t\t" << N * (T - 1)
-                    / (simdpacktimewm) << "\t\t" << N * (T - 1)
-                    / (simdunpacktime) << "\t\t";
+                    / double(simdpacktime) << "\t\t" << N * (T - 1)
+                    / double(simdpacktimewm) << "\t\t" << N * (T - 1)
+                    / double(simdunpacktime) << "\t\t";
             cout<< N * (T - 1)
-                            / (horizontalunpacktime) << "\t\t";
+                            / double(horizontalunpacktime) << "\t\t";
 
 
             cout << endl;
